Add link to web IRC to reduce friction for new users
Users seeking assistance may not know how to use IRC but also might not
want to have to sign up for Discord just to ask a question. To fix this,
add a link to the channel in Libera's webchat so users can start
chatting in just two clicks. Users familiar with IRC can use the ircs://
link titled "connect with your own client" to connect with their own IRC
client.